requirements.txt.txt
streamlit
pandas
numpy
matplotlib
pillow